Refactor: Move _DATA_DIR to data_dir.py.
