+++ tests/functional/scripts/pyi_lib_PyQt4-uic.py
@@ -24,7 +24,7 @@ from PyQt4.QtCore import QTimer
 
 # Local imports
 # -------------
-from tests.functional.get_data_dir import get_data_dir
+from tests.functional.data_dir import get_data_dir
 
 # Test code
 # ---------
